+++ CHANGELOG.md
@@ -1,8 +1,11 @@
+## 0.4.0
+support android-gradle-plugin 3.6.x & support R.jar
+
 ## 0.3.0
 fix ApkTool de-compile issues & remove symbol output checks
 
 ## 0.2.0
-add support android-gradle-plugin 3.0+
+add support android-gradle-plugin 3.0.x
 
 ## 0.1.0
 first release